Consolidated MyEditor and ClinicalNotes
