+++ src/notes/ClinicalNotes.css
@@ -21,3 +21,89 @@
 #data {
     color: black;
 }
+
+.MyEditor-root {
+    font-family: 'Georgia', serif;
+}
+
+.MyEditor-root .toolbar-menu > * + * {
+    margin-left: 15px;
+}
+
+.MyEditor-root .toolbar-menu > * {
+    display: inline-block;
+}
+
+.MyEditor-root .toolbar-menu {
+    padding: 1px 0 17px 5px;
+    margin-bottom: 20px;
+    border-bottom: 1px solid #ddd;
+}
+
+.MyEditor-root .toolbar-menu .button {
+    color: #ccc;
+    cursor: pointer;
+}
+
+.MyEditor-root .toolbar-menu .button[data-active="true"] {
+  color: black;
+}
+
+.MyEditor-root .autocomplete-menu { 
+    position: absolute;
+    box-sizing: border-box;
+    border-radius: 2px;
+    -webkit-tap-highlight-color: rgba(0, 0, 0, 0);
+    box-shadow: rgba(0, 0, 0, 0.12) 0px 1px 6px, rgba(0, 0, 0, 0.12) 0px 1px 4px;
+    transition: all 450ms cubic-bezier(0.23, 1, 0.32, 1) 0ms;
+    background-color: rgb(255, 255, 255);
+    z-index: 1111;
+}
+
+.MyEditor-root .autocomplete-menu > * {
+    display: block;
+    padding: 7px 10px;
+}
+
+.MyEditor-root .autocomplete-menu .menu-item[data-active="true"] {
+    background-color: rgba(0,0,0,0.1);
+}
+
+
+.MyEditor-root .shorthand-menu { 
+    position: absolute;
+    box-sizing: border-box;
+    border-radius: 2px;
+    -webkit-tap-highlight-color: rgba(0, 0, 0, 0);
+    box-shadow: rgba(0, 0, 0, 0.12) 0px 1px 6px, rgba(0, 0, 0, 0.12) 0px 1px 4px;
+    transition: all 450ms cubic-bezier(0.23, 1, 0.32, 1) 0ms;
+    background-color: rgb(255, 255, 255);
+    z-index: 1111;
+}
+
+.MyEditor-root .shorthand-menu > * {
+    display: block;
+    padding: 7px 10px;
+}
+
+.MyEditor-root .shorthand-menu .menu-item[data-active="true"] {
+    background-color: rgba(0,0,0,0.1);
+}
+
+
+.MyEditor-root p {
+    display: inline;
+    margin:0;
+}
+
+.MyEditor-root span {
+    display: inline;
+}
+
+.clear{
+  clear: both;
+  display: block;  
+  content: "";
+  margin: 10px 0;
+  width: 100%;  
+}
